# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CollAPI - API de Cobranzas",
        "description": "API REST para el sistema de gestión de cobranzas",
        "contact": {
            "email": "soporte@collapi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/contacts": {
            "post": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Crear contacto",
                "description": "Crea un nuevo contacto asociado a un cliente",
                "operationId": "storeContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "phone_type",
                                    "client_id"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "description": "Número de teléfono",
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "0991234567"
                                    },
                                    "phone_type": {
                                        "description": "Tipo de teléfono (CELULAR, FIJO, etc.)",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "CELULAR"
                                    },
                                    "client_id": {
                                        "description": "ID del cliente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Nombre del contacto",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan Pérez"
                                    },
                                    "phone_status": {
                                        "description": "Estado del teléfono",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "ACTIVE"
                                    },
                                    "calls_effective": {
                                        "description": "Cantidad de llamadas efectivas",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "calls_not_effective": {
                                        "description": "Cantidad de llamadas no efectivas",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "credit_id": {
                                        "description": "ID del crédito asociado (opcional)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto creado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "phone_type": {
                                                    "type": "string"
                                                },
                                                "phone_status": {
                                                    "type": "string"
                                                },
                                                "calls_effective": {
                                                    "type": "integer"
                                                },
                                                "calls_not_effective": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "integer"
                                                },
                                                "credit_id": {
                                                    "type": "integer"
                                                },
                                                "created_by": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "properties": {
                                                "phone_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El número de teléfono es obligatorio."
                                                    }
                                                },
                                                "phone_type": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El tipo de teléfono es obligatorio."
                                                    }
                                                },
                                                "client_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El cliente es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contacts/{id}": {
            "put": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Actualizar contacto",
                "description": "Actualiza un contacto existente",
                "operationId": "updateContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del contacto",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan Perez"
                                    },
                                    "phone_number": {
                                        "description": "Numero de telefono",
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "0991234567"
                                    },
                                    "phone_type": {
                                        "description": "Tipo de telefono",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "CELULAR"
                                    },
                                    "phone_status": {
                                        "description": "Estado del telefono",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "ACTIVE"
                                    },
                                    "calls_effective": {
                                        "description": "Llamadas efectivas",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "calls_not_effective": {
                                        "description": "Llamadas no efectivas",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "client_id": {
                                        "description": "ID del cliente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "credit_id": {
                                        "description": "ID del credito",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contacto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto actualizado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "phone_type": {
                                                    "type": "string"
                                                },
                                                "phone_status": {
                                                    "type": "string"
                                                },
                                                "calls_effective": {
                                                    "type": "integer"
                                                },
                                                "calls_not_effective": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "integer"
                                                },
                                                "client_name": {
                                                    "type": "string"
                                                },
                                                "client_ci": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}